@@ -3,19 +3,19 @@ import 'package:flutter/material.dart';
 /// [Widget] for rendering the AM/PM button
 class AmPm extends StatelessWidget {
   /// Currently selected by user
-  final String selected;
+  final String? selected;
 
   /// [onChange] handler for AM/PM
-  final void Function(String) onChange;
+  final void Function(String)? onChange;
 
   /// Accent color to be used for the button
-  final Color accentColor;
+  final Color? accentColor;
 
   /// Accent color to be used for the unselected button
-  final Color unselectedColor;
+  final Color? unselectedColor;
 
   /// Default [TextStyle]
-  final _style = TextStyle(fontSize: 20);
+  final _style = const TextStyle(fontSize: 20);
 
   /// Initialize the buttons
   AmPm({this.selected, this.onChange, this.accentColor, this.unselectedColor});
@@ -23,7 +23,7 @@ class AmPm extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final isAm = selected == 'am';
-    final unselectedOpacity = 0.5;
+    const unselectedOpacity = 0.5;
 
     return Container(
       child: Row(
@@ -34,7 +34,7 @@ class AmPm extends StatelessWidget {
             child: InkWell(
               onTap: !isAm
                   ? () {
-                      onChange("am");
+                      onChange!("am");
                     }
                   : null,
               child: Padding(
@@ -58,7 +58,7 @@ class AmPm extends StatelessWidget {
             child: InkWell(
               onTap: isAm
                   ? () {
-                      onChange("pm");
+                      onChange!("pm");
                     }
                   : null,
               child: Padding(
